changed a few spelling errors, and updated and expanded MMU

+++ src/PhysicalMemory.java
@@ -5,12 +5,12 @@ public class PhysicalMemory{
     so...
     0x ABC      Page number==>[A][BC]<==Offset
     */
-    private double ram[][] = new double[16][256];
+    private int ram[][] = new int[16][256];
 
     public double getPhysicalMem(int row, int col){
         return ram[row][col];
     }
-    public void setPhysicalMem(int row, int col, double val){
+    public void setPhysicalMem(int row, int col, int val){
         ram[row][col]= val;
     }
 }
